--- a/app/models/principal.rb
+++ b/app/models/principal.rb
@@ -66,16 +66,11 @@ class Principal < ActiveRecord::Base
   scope :active_or_registered_like, ->(query) { active_or_registered.like(query) }
 
   scope :in_project, ->(project) {
-    projects = Array(project)
-    subquery = "SELECT DISTINCT user_id FROM members WHERE project_id IN (?)"
-    condition = ["#{Principal.table_name}.id IN (#{subquery})",
-                 projects.map(&:id)]
-
-    where(condition)
+    where(id: Member.of(project).select(:user_id))
   }
 
   scope :not_in_project, ->(project) {
-    where("id NOT IN (select m.user_id FROM members as m where m.project_id = #{project.id})")
+    where.not(id: Member.of(project).select(:user_id))
   }
 
   # Active non-anonymous principals scope