--- a/docs/api/apiv3-documentation.apib
+++ b/docs/api/apiv3-documentation.apib
@@ -13,6 +13,7 @@ FORMAT: 1A
 <!-- include(apiv3/endpoints/custom-actions.apib) -->
 <!-- include(apiv3/endpoints/custom-options.apib) -->
 <!-- include(apiv3/endpoints/forms.apib) -->
+<!-- include(apiv3/endpoints/groups.apib) -->
 <!-- include(apiv3/endpoints/help_texts.apib) -->
 <!-- include(apiv3/endpoints/principals.apib) -->
 <!-- include(apiv3/endpoints/priorities.apib) -->