--- a/docs/api/apiv3/endpoints/users.apib
+++ b/docs/api/apiv3/endpoints/users.apib
@@ -110,7 +110,7 @@ Please note that custom fields are not yet supported by the api although the bac
 
     Returned if the user does not exist or if the API user does not have permission to view them.
 
-    **Required permission** The user needs to be locked in if the installation is configured to pervent anonymous access
+    **Required permission** The user needs to be locked in if the installation is configured to prevent anonymous access
 
     + Body