--- a/frontend/app/components/common/autocomplete/auto-complete-helper.service.ts
+++ b/frontend/app/components/common/autocomplete/auto-complete-helper.service.ts
@@ -36,7 +36,7 @@ export class AutoCompleteHelperService {
       at: at,
       startWithSpace: true,
       searchKey: 'id_principal',
-      displayTpl: '<li data-value="#{subtype}#${id}">${name}</li>',
+      displayTpl: '<li data-value="#{_type}#${id}">${name}</li>',
       insertTpl: "${typePrefix}#${id}",
       limit: 10,
       highlightFirst: true,
@@ -55,7 +55,7 @@ export class AutoCompleteHelperService {
                 const principals = data["_embedded"]["elements"];
                 for (let i = principals.length - 1; i >= 0; i--) {
                   principals[i]['id_principal'] = principals[i]['id'].toString() + ' ' + principals[i]['name'];
-                  principals[i]['typePrefix'] = principals[i]['subtype'].toLowerCase();
+                  principals[i]['typePrefix'] = principals[i]['_type'].toLowerCase();
                 }
 
                 if (angular.element(textarea).is(':visible')) {