--- a/lib/api/v3/principals/associated_subclass_lambda.rb
+++ b/lib/api/v3/principals/associated_subclass_lambda.rb
@@ -1,4 +1,5 @@
 #-- encoding: UTF-8
+
 #-- copyright
 # OpenProject is a project management system.
 # Copyright (C) 2012-2018 the OpenProject Foundation (OPF)
@@ -29,9 +30,24 @@
 
 module API
   module V3
-    module Users
-      class UserCollectionRepresenter < ::API::Decorators::UnpaginatedCollection
-        include API::V3::Principals::GroupOrUserElements
+    module Principals
+      module AssociatedSubclassLambda
+        def self.getter(name)
+          ->(*) {
+            instance = represented.send(name)
+
+            case instance
+            when User
+              ::API::V3::Users::UserRepresenter.new(represented.send(name), current_user: current_user)
+            when Group
+              ::API::V3::Groups::GroupRepresenter.new(represented.send(name), current_user: current_user)
+            when NilClass
+              nil
+            else
+              raise "undefined subclass for #{instance}"
+            end
+          }
+        end
       end
     end
   end