--- a/lib/api/v3/priorities/priority_representer.rb
+++ b/lib/api/v3/priorities/priority_representer.rb
@@ -34,6 +34,8 @@ module API
   module V3
     module Priorities
       class PriorityRepresenter < ::API::Decorators::Single
+        include API::Caching::CachedRepresenter
+
         self_link
 
         property :id, render_nil: true