… frontend/app/templates/components/user_field.html
… lib/api/v3/principals/group_or_user_elements.rb
… lib/api/v3/principals/principal_representer.rb
… lib/api/v3/principals/principals_api.rb
… lib/api/v3/projects/project_representer.rb
… lib/api/v3/root.rb
… lib/api/v3/statuses/status_representer.rb
… lib/api/v3/types/type_representer.rb
… lib/api/v3/users/paginated_user_collection_representer.rb
… lib/api/v3/users/user_representer.rb
… lib/api/v3/utilities/path_helper.rb
… lib/api/v3/versions/version_representer.rb
… lib/api/v3/work_packages/schema/specific_work_package_schema.rb
… lib/api/v3/work_packages/schema/typed_work_package_schema.rb
… lib/api/v3/work_packages/schema/work_package_schema_representer.rb
… lib/api/v3/work_packages/schema/work_package_schemas_api.rb
… lib/api/v3/work_packages/work_package_payload_representer.rb
… lib/api/v3/work_packages/work_package_representer.rb
… lib/open_project/cache.rb
… lib/open_project/cache/cache_key.rb
… lib/open_project/plugins/acts_as_op_engine.rb
… spec/factories/group_factory.rb
… spec/factories/principal_factory.rb
… spec/factories/user_factory.rb
… spec/features/users/create_spec.rb
… spec/features/work_packages/custom_actions_spec.rb
… spec/lib/api/v3/attachments/attachment_representer_spec.rb
… spec/lib/api/v3/categories/category_representer_spec.rb
… spec/lib/api/v3/groups/group_representer_spec.rb
… spec/lib/api/v3/priorities/priority_representer_spec.rb