https://community.openproject.com/projects/openproject/work_packages/27541

### TODO

- [x] project representer
- [x] user representer
- [x] priority representer 
- [x] status representer 
- [x] attachments representer 
- [x] version representer
- [x] type representer  
- [x] category representer 
- [x] work_package schema representer 
- [x] work package representer
- ~~work package eager loading~~ will be extracted to a separate PR
- [x] adapt plugins modifying representers
  - [x] costs (https://github.com/finnlabs/openproject-costs/pull/307)
  - [x] backlogs (https://github.com/finnlabs/openproject-backlogs/pull/265)
- [x] remove `subtype` dependencies in the front end as there is now a `Group` resource

### Preliminary results

Setting: 
* OpenProject Core only

#### Without caching:
```
> ab -n 300 -H "Authorization: Basic YXBpa2V5OjVlZTM2ZjdlNzNmOTFkYjcyOWNkOWNhZWYwYTRhMjEyMGNlYzg2NGM4YjI5YTNlZGZjZGVjMzk0NDhiZTcxYTQ=" http://localhost:3000/api/v3/work_packages?pageSize=50&offset=1
[1] 7143
This is ApacheBench, Version 2.3 <$Revision: 1807734 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient)
Completed 100 requests
Completed 200 requests
Completed 300 requests
Finished 300 requests


Server Software:        thin
Server Hostname:        localhost
Server Port:            3000

Document Path:          /api/v3/work_packages?pageSize=50
Document Length:        325220 bytes

Concurrency Level:      1
Time taken for tests:   533.943 seconds
Complete requests:      300
Failed requests:        0
Total transferred:      97800600 bytes
HTML transferred:       97566000 bytes
Requests per second:    0.56 [#/sec] (mean)
Time per request:       1779.812 [ms] (mean)
Time per request:       1779.812 [ms] (mean, across all concurrent requests)
Transfer rate:          178.87 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.1      0       1
Processing:  1529 1780 294.7   1706    5005
Waiting:     1529 1779 294.7   1706    5005
Total:       1529 1780 294.7   1706    5006

Percentage of the requests served within a certain time (ms)
  50%   1706
  66%   1769
  75%   1835
  80%   1884
  90%   2002
  95%   2255
  98%   2523
  99%   3247
 100%   5006 (longest request)
```

#### With caching:
```
> ab -n 300 -H "Authorization: Basic YXBpa2V5OjVlZTM2ZjdlNzNmOTFkYjcyOWNkOWNhZWYwYTRhMjEyMGNlYzg2NGM4YjI5YTNlZGZjZGVjMzk0NDhiZTcxYTQ=" http://localhost:3000/api/v3/work_packages?pageSize=50&offset=1
[1] 6194
This is ApacheBench, Version 2.3 <$Revision: 1807734 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
  Licensed to The Apache Software Foundation, http://www.apache.org/

  Benchmarking localhost (be patient)
Completed 100 requests
Completed 200 requests
Completed 300 requests
Finished 300 requests


Server Software:        thin
Server Hostname:        localhost
Server Port:            3000

Document Path:          /api/v3/work_packages?pageSize=50
Document Length:        325220 bytes

Concurrency Level:      1
Time taken for tests:   291.349 seconds
Complete requests:      300
Failed requests:        0
Total transferred:      97800600 bytes
HTML transferred:       97566000 bytes
Requests per second:    1.03 [#/sec] (mean)
Time per request:       971.165 [ms] (mean)
Time per request:       971.165 [ms] (mean, across all concurrent requests)
Transfer rate:          327.81 [Kbytes/sec] received

Connection Times (ms)
min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:   859  971  75.9    933    1179
Waiting:      858  971  75.9    933    1178
Total:        859  971  75.9    934    1179

Percentage of the requests served within a certain time (ms)
  50%    934
  66%   1018
  75%   1058
  80%   1065
  90%   1078
  95%   1092
  98%   1108
  99%   1124
 100%   1179 (longest request)
```

#### Analysis
Although eager loading has not yet been implemented, caching the representers shows a significant improvement in speed. Response time drops to about 60% with a lower standard deviation. 

### Background

Improves on the [first POC of caching API representers](https://github.com/opf/openproject/pull/5086).

As of now, the improvements in response time need to be taken with a grain of salt but the preliminary measurements indicate a reduction of the response time for `api/v3/work_packages` to about 50 - 33%. E.g. a request for 50 work packages takes about 1s on my machine. With the changes of this PR applied, the response is served in about 350ms. For 100 work package the response time drops from 1.5 seconds to 400ms.

The first POC's approach was to distinguish the representer into layers of concerns which where individually cached. Upon a request, the cache key, mainly derived from the user's permissions would determine the separate layers that where then combined into the final response body.

This PR takes the opposite approach. The representer is rendered without any consideration for the current user's permission which results in a representation fit for responding to an admin's request. For non admins, that cached representation is stripped of those parts not fit for the requesting user's permissions. As an example, the project representer has the `createWorkPackages` link:

```
{
    "_type": "Project",
    ...
    "_links": {
        "self": {
            "href": "/api/v3/projects/1",
            "title": "Seeded Project"
        },
        "createWorkPackage": {
            "href": "/api/v3/projects/1/work_packages/form",
            "method": "post"
        },
        ...
    }
}
```

The cached representation will contain that link and thus be able to satisfy the complete permissions an admin has. But when a user lacking the `add_work_packages` permission requests that same project, the same representation will be the used as the foundation for the response, but before returning the response the application removes the `createWorkPackage` link.

Compared to the first POC this approach has a couple of advantages:
* High probability of a cache hit as the same result can be used as the base for all users' requests. In effect, that should mean to only need to do the heavy lifting of determining the response once for every saved state of a resource. E.g. the project's representation can be reused for all users' requests until the project is saved again.
* Simple to implement (see commits)

However, rending the representer is only responsible for the lesser part of time spent when responding to `/api/v3/work_packages/` (with our without project scope). That type of request is responsible for over 30% of computational resources spent on SaaS.  

![image](https://user-images.githubusercontent.com/617519/36245673-74448e72-122c-11e8-9070-73a236e8cb10.png)

The bulk of time is spent on instantiating the AR models after fetching them from the DB.
 
![image](https://user-images.githubusercontent.com/617519/36245746-d80b3262-122c-11e8-8d85-3128a9de38b1.png)

 In the past, we optimized the response time by eager loading all of the models before rendering. But as the work package representer contains information from a lot of different models (`WorkPackage`, `User` (3 times), `Status`, ...), that approach has lead to the application having to instantiating a lot of models which is quite costly. 

To maximize the gains received from caching, we also need to reduce the amount of models instantiated. But as the cached representation needs to be invalidated once any of the models change (e.g. when the status is renamed, the cached work package representation is no longer valid). 

Invalidation should happen by changing the cache key. The cache key will typically consist of a class, the id and a timestamp. For our purposes, we need to extend that to also include the timestamps of all the rendered models. As Rails' eager loading mechanism is not able to select only a few fields we need to write custom queries of the form:
```
WorkPackages.joins(:author).select('work_packages.*, users.updated_on)
```

This approach to caching has only one drawback, which is that if a lot of resources in a collection are uncached, that response is going to take a lot more time as the models are than fetched with N+1 queries. This effect can be avoided in part by preemptively loading a large set of representations into the cache by e.g. a rake task.