@oliverguenther 2018-04-26
Clicked through the CE edition and LGTM. Since you said we should improve caching to ideally accessing it only once, do note that we don't have shared memcached instances for all packaged installations atm. That means that this will happen on every application server. To mitigate this, we will need:

1. A way to populate memcached server groups
2. A way to split memcached server groups (think dtag network-separated application servers where we won't be able to share the same cache)

Can you open the respective tickets if you think this is required before deployment?